% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-redcap.R
\name{read_redcap_raw}
\alias{read_redcap_raw}
\title{Read raw REDCap export}
\usage{
read_redcap_raw(config, fields)
}
\arguments{
\item{config}{A \code{dashboard_config} object with REDCap credentials.}

\item{fields}{Character vector of REDCap field names to export.}
}
\value{
The raw (long, sparse) REDCap export as a data frame.
}
\description{
Read raw REDCap export
}
\keyword{internal}
