% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-redcap.R
\name{redcap_audit_fields}
\alias{redcap_audit_fields}
\title{REDCap fields exported for the dashboard}
\usage{
redcap_audit_fields()
}
\value{
Character vector of REDCap field names.
}
\description{
REDCap fields exported for the dashboard
}
\keyword{internal}
